modules/locations_grpc_kafka_producer/requirements.txt
grpcio==1.56.2
kafka_python==2.0.2
protobuf==4.23.4
